Move patch back to scc file
Following on from comments by Bruce Ashfield, BSP patches should be
applied via the scc file and not using SRC_URI.

He also suggested putting the "include ktypes" line back, but with
a specific option to ensure patches are not re-applied. The advantage
of this is that we then get the standard kernel config options applied.